front-end/front-office-mobile/appshareworld/test/SignUp/SignUp.spec.tsx
import renderer from 'react-test-renderer';
import SignUpScreen from '../../src/screens/authScreen/SignUp';
import React from 'react';



let findTextInput = function(tree: any, element: any){
   return true;
    }



    it('find TextInput', () => {
        const tree = renderer.create(<SignUpScreen/>).toJSON();
        expect(findTextInput(tree, 'lastname')).toBeDefined();
    })
    it('find TextInput', () => {
        const tree = renderer.create(<SignUpScreen/>).toJSON();
        expect(findTextInput(tree, 'firstname')).toBeDefined();
    })
    it('find TextInput', () => {
        const tree = renderer.create(<SignUpScreen/>).toJSON();
        expect(findTextInput(tree, 'phone')).toBeDefined();
    })
    it('find TextInput', () => {
        const tree = renderer.create(<SignUpScreen/>).toJSON();
        expect(findTextInput(tree, 'avatar')).toBeDefined();
    })
    it('find TextInput', () => {
        const tree = renderer.create(<SignUpScreen/>).toJSON();
        expect(findTextInput(tree, 'email')).toBeDefined();
    })
    it('find TextInput', () => {
        const tree = renderer.create(<SignUpScreen/>).toJSON();
        expect(findTextInput(tree, 'password')).toBeDefined();
    })
    it('find TextInput', () => {
        const tree = renderer.create(<SignUpScreen/>).toJSON();
        expect(findTextInput(tree, 'verifypassword')).toBeDefined();
    })
    it('find TextInput', () => {
        const tree = renderer.create(<SignUpScreen/>).toJSON();
        expect(findTextInput(tree, 'number')).toBeDefined();
    })
    it('find TextInput', () => {
        const tree = renderer.create(<SignUpScreen/>).toJSON();
        expect(findTextInput(tree, 'street')).toBeDefined();
    })
    it('find TextInput', () => {
        const tree = renderer.create(<SignUpScreen/>).toJSON();
        expect(findTextInput(tree, 'addressmore')).toBeDefined();
    })
    it('find TextInput', () => {
        const tree = renderer.create(<SignUpScreen/>).toJSON();
        expect(findTextInput(tree, 'city')).toBeDefined();
    })
    it('find TextInput', () => {
        const tree = renderer.create(<SignUpScreen/>).toJSON();
        expect(findTextInput(tree, 'zipcode')).toBeDefined();
    })
    it('find  Submit', () => {
        const tree = renderer.create(<SignUpScreen/>).toJSON();
        expect(findTextInput(tree, 'submit')).toBeDefined();
    })
    // it('Renders snapshot as expected', () => {
    //     const tree = renderer.create(<SignUpScreen />).toJSON();
    //     expect(tree).toMatchSnapshot();
    // });

   